api_gateway/requirements.txt
Flask==2.2.5
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
pyjwt==2.8.0
bcrypt==4.1.2
sqlalchemy==2.0.23
pydantic==2.5.2
psycopg2-binary==2.9.9  # For PostgreSQL